SIMPLE CRUD

+++ app/controllers/likes_controller.rb
@@ -1,10 +1,11 @@
 class LikesController < ApplicationController
+    before_action :authenticate_user!
     
     def create
         idea = Idea.find params[:idea_id]
         like = Like.new(user: current_user, idea: idea)
         if like.save
-            flash[:success] = "I Like idea"
+            flash[:success] = "I Like the idea"
             redirect_to idea
         else
             flash[:danger] = like.errors.full_messages.join(", ")
@@ -14,8 +15,12 @@ class LikesController < ApplicationController
 
     def destroy
         like = Like.find params[:id]
-        like.destroy
+        if can? like.destroy,like
+            like.destroy
         flash[:success] = "Unlike the idea"
         redirect_to like.idea
+        else
+            redirect_to like.idea, alert: "can't delete like"
+        end
     end
 end
